Move QScriptEvaluator to connector.evaluate, remove from HeavyweightDatsource
